requirements.txt
events 
pybluez 
pyqt5 
pyserial 
pyudev 
tqdm  